Refactor EmailClient and Validation classes into individual module files.
