+++ src/utils/validation.mjs
@@ -1,0 +1,46 @@
+/*
+ * Copyright (C) 2025 brittni watkins.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+export class Validation {
+    static isValidString(input) {
+        const validType = typeof input === 'string';
+        let validContent = false;
+
+        if (validType) {
+            validContent = input.trim().length > 0;
+        }
+
+        return validType && validContent;
+    }
+
+    static isValidNumber(input) {
+        return (typeof input === 'number') && !isNaN(input);
+    }
+
+    static sanitizeString(input) {
+        if (Validation.isValidString(input)) {
+            return input.trim();
+        } else {
+            return undefined;
+        }
+    }
+}
